refactor: Move JsTaskRunner into its own directory
