+++ packages/@n8n/task-runner/src/js-task-runner/obj-utils.ts
